auth, user mobility pass and pricing services tested et ca marche!

+++ user-service/src/main/java/com/smartmobility/userservice/exception/SoldeInsuffisantException.java
@@ -1,9 +1,10 @@
 package com.smartmobility.userservice.exception;
 
+import java.math.BigDecimal;
 
 public class SoldeInsuffisantException extends RuntimeException {
 
-    public SoldeInsuffisantException(Double soldeActuel, Double montantDemande) {
+    public SoldeInsuffisantException(BigDecimal soldeActuel, BigDecimal montantDemande) {
         super("Solde insuffisant. Solde actuel : " + soldeActuel + " FCFA, montant demandé : " + montantDemande + " FCFA");
     }
 }
